+++ @types-extensions/css/index.d.ts
@@ -2,3 +2,7 @@ declare module '*.mod.css' {
   const classes: { [key: string]: string };
   export default classes;
 }
+declare module '*.module.css' {
+  const classes: { [key: string]: string };
+  export default classes;
+}