+++ application-templates/starter/src/components/entry-point/entry-point.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import { lazy } from 'react';
 import {
   ApplicationShell,
   setupGlobalErrorListener,
@@ -8,7 +8,7 @@ import loadMessages from '../../load-messages';
 // Here we split up the main (app) bundle with the actual application business logic.
 // Splitting by route is usually recommended and you can potentially have a splitting
 // point for each route. More info at https://reactjs.org/docs/code-splitting.html
-const AsyncApplicationRoutes = React.lazy(() =>
+const AsyncApplicationRoutes = lazy(() =>
   import('../../routes' /* webpackChunkName: "routes" */)
 );