+++ packages/react-notifications/src/components/notifications-list/index.ts
@@ -1,2 +1,1 @@
 export { default } from './notifications-list';
-export { selectPageNotifications } from './selectors';