+++ packages/application-shell/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @commercetools-frontend/application-shell
 
+## 21.4.0
+
+### Minor Changes
+
+- [#2572](https://github.com/commercetools/merchant-center-application-kit/pull/2572) [`5b06e97a`](https://github.com/commercetools/merchant-center-application-kit/commit/5b06e97a65b48255d45f4d66acb6d52548025cc4) Thanks [@emmenko](https://github.com/emmenko)! - Bump uikit packages to v15.
+
+* [#2550](https://github.com/commercetools/merchant-center-application-kit/pull/2550) [`1065e661`](https://github.com/commercetools/merchant-center-application-kit/commit/1065e661f4bd7bd9b6096fdc6296c0341129f301) Thanks [@kark](https://github.com/kark)! - Add a new `<SuspendedRoute>` component.
+
+  This component is exactly like the `<Route>` component of `react-router` but it wraps the children with `React.Suspense`, allowing the children to be dynamically loaded with `React.lazy`. This can be used for code splitting components at the route level.
+
+  ```jsx
+  import { Switch } from 'react-router-dom';
+  import { lazy } from 'react';
+
+  const LazyLoadedComponent = lazy(() => import('./lazy-loaded-component'));
+
+  <Switch>
+    <SuspendedRoute>
+      <LazyLoadedComponent />
+    </SuspendedRoute>
+  </Switch>;
+  ```
+
+### Patch Changes
+
+- [#2563](https://github.com/commercetools/merchant-center-application-kit/pull/2563) [`238bd34a`](https://github.com/commercetools/merchant-center-application-kit/commit/238bd34a6045fd4cda443585b714ac0689ecd335) Thanks [@kark](https://github.com/kark)! - Update `react-intl` to version `^5.25.0`
+
+- Updated dependencies [[`238bd34a`](https://github.com/commercetools/merchant-center-application-kit/commit/238bd34a6045fd4cda443585b714ac0689ecd335), [`5b06e97a`](https://github.com/commercetools/merchant-center-application-kit/commit/5b06e97a65b48255d45f4d66acb6d52548025cc4)]:
+  - @commercetools-frontend/application-components@21.4.0
+  - @commercetools-frontend/i18n@21.4.0
+  - @commercetools-frontend/react-notifications@21.4.0
+
+## 21.3.5
+
+### Patch Changes
+
+- [#2554](https://github.com/commercetools/merchant-center-application-kit/pull/2554) [`8e170c65`](https://github.com/commercetools/merchant-center-application-kit/commit/8e170c6574e8d52738b3a00022fedc8df03ce048) Thanks [@emmenko](https://github.com/emmenko)! - Fix offset calculation for modal pages when notifications are opened.
+
+* [#2543](https://github.com/commercetools/merchant-center-application-kit/pull/2543) [`ab3cbd23`](https://github.com/commercetools/merchant-center-application-kit/commit/ab3cbd2386fb14b6dd25556262d00da71fc78588) Thanks [@Rhotimee](https://github.com/Rhotimee)! - Add tracking for some nav menu items (Support link, expand toggle).
+
+  Fix passing `onMenuItemClick` prop to the nav menu.
+
+* Updated dependencies [[`8e170c65`](https://github.com/commercetools/merchant-center-application-kit/commit/8e170c6574e8d52738b3a00022fedc8df03ce048)]:
+  - @commercetools-frontend/application-components@21.3.5
+  - @commercetools-frontend/react-notifications@21.3.5
+
 ## 21.3.4
 
 ### Patch Changes